@@ -1,20 +1,20 @@
 {
-    "name": "nightdiraven/amari",
-    "description": "Some monkey translation service and json field for laravel. Do not install it. I'm serious",
-    "authors": [
-        {
-            "name": "Ivan Dolgov",
-            "email": "nightdiraven@gmail.com"
-        }
-    ],
-    "license": "MIT",
-    "minimum-stability": "dev",
-    "require": {
-        "php": ">=5.3.0"
-    },
-    "autoload": {
-        "psr-0": {
-            "Amari": "src/"
-        }
+  "name": "nightdiraven/amari",
+  "description": "Some monkey translation service and json field for laravel. Do not install it. I'm serious",
+  "authors": [
+    {
+      "name": "Ivan Dolgov",
+      "email": "nightdiraven@gmail.com"
     }
+  ],
+  "license": "MIT",
+  "minimum-stability": "dev",
+  "require": {
+    "php": ">=7.0"
+  },
+  "autoload": {
+    "psr-0": {
+      "Amari": "src/"
+    }
+  }
 }
